Add umask parameter
My local security policy dictates that the system's default umask has to
be 027 instead of the normal 022.  This meant that CPAN modules being
installed by this puppet module were only readable by root.

This commit adds a new `umask` parameter than can be used to force the
use of a specific umask when creating or updating CPAN modules.