Fix ensure with nonstandard versions
This increases support for DragonFly BSD as well as FreeBSD, making
them first class citizens for this module. Additionally, it creates a
new structured fact called perl, which contains the various version
numbers that can be referenced for what version of Perl is the system
default. It was mainly added to properly support the config paths for
the default @INC on DragonFly and FreeBSD. The cpan provider was
updated to fix a few bugs around how it handled version strings, and
using rubocop, some additional cleanup was done while there.